fixed some gcc compiler warnings for jtorch.  jtorch and jtorch_test now compile (but the test fails).

+++ jtorch/src/jtorch/join_table.cpp
@@ -55,7 +55,7 @@ namespace jtorch {
       size += in(i)->dataSize();
     }
      
-    if (output != NULL && size != output->dataSize()) {
+    if (output != NULL && size != static_cast<int>(output->dataSize())) {
       SAFE_DELETE(output);
     }
